requirements.txt
Django~=5.1.2
django-allauth~=65.1.0
django-crispy-forms~=2.3
numpy~=2.1.2
websockets~=13.1
pyfingerprint~=1.5
cryptography~=43.0.3